Bug #835952: pep8 failures do not cause the tests to fail
Add set -eu to run_tests.sh.  This will cause it to fail whenever anything
goes wrong, which is exactly what we want in a test script.

To do this, I had to remove the use of the "let" keyword, which has a bizarre
exit status in bash.

I also removed the "|| exit" after run_tests, which means that this script
will now exit with status 1, not status 0, if run_tests fails.