Fixes an issue where 'invalid literal for int' would occur when listing images after making a v1.1 server snapshot (with a UUID).
v1.1 image id's are now treated as strings (not integer ID's). The v1.0
API still tries to treat image id's as integers but doesn't fail
miserably if they are uuid's either.

This should pave the way for image ID's as uuids and more closely matches
the v1.1 spec with regards to images and the server refs they contain.