Import copies of the required openstack.common modules
The APIs melange is using are "incubating" in openstack-common. See:

  http://wiki.openstack.org/CommonLibrary#Incubation

A config file is included for openstack-common's copy-and-paste script
to make it easy to update the code e.g.

  $> git clone .../melange
  $> git clone .../openstack-common
  $> cd openstack-common
  $> python update.py ../melange/

Note: the copies of the openstack.common modules should never be directly
modified, instead you should always make the modifications in openstack-common
and copy them across using the script.

Change-Id: I0245e30e4b4a9a2214c4f8a291b7f2caa54ab97e